requirements.txt
pillow==10.4.0
tqdm
easydict==1.13
opencv-python-headless==4.11.0.86
open3d==0.19.0
scipy==1.14.1
rembg==2.0.60
onnxruntime==1.20.1
trimesh
xatlas==0.0.9
pyvista==0.44.2
pymeshfix==0.17.0
igraph==0.11.8
datasets==4.3.0
deepspeed
ipdb
modelcards
omegaconf
einops
imageio
imageio-ffmpeg
ftfy
wandb
pandas
diffusers[torch]==0.33.1
transformers>=4.51.0
bitsandbytes
decord==0.6.0
accelerate
tensorboard
modelcards
xformers==0.0.20
git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
ipython==8.12.0
av
pynvml
lpips
unitest
scikit-video
tryo
jaxtyping
open-clip-torch
